Introduce Settings class. Not used everywhere yet
